+++ nebula/core/aggregation/updatehandlers/dflupdatehandler.py
@@ -40,6 +40,10 @@ class DFLUpdateHandler(UpdateHandler):
        self._sources_received = set()
        self._round_updates_lock = Locker(name="round_updates_lock", async_lock=True) # se coge cuando se empieza a comprobar si estan todas las updates
        self._update_federation_lock = Locker(name="update_federation_lock", async_lock=True)
+       self._notification_sent_lock =  Locker(name="notification_sent_lock", async_lock=True)
+       self._notification = False
+       self._missing_ones = set()
+       self._nodes_using_historic = set()
        
     @property
     def us(self):
@@ -50,15 +54,6 @@ class DFLUpdateHandler(UpdateHandler):
         return self._aggregator  
        
     async def round_expected_updates(self, federation_nodes: set):
-        """
-            Initializes the expected updates for the current round.
-
-            This method updates the list of expected sources (`_sources_expected`) for the current training round 
-            and ensures that their respective update storage is initialized if they were not previously registered.
-
-            Args:
-                federation_nodes (set): A set of node identifiers expected to provide updates in the current round.
-        """
         await self._update_federation_lock.acquire_async()
         await self._updates_storage_lock.acquire_async()
         self._sources_expected = federation_nodes.copy()
@@ -83,6 +78,8 @@ class DFLUpdateHandler(UpdateHandler):
         # Lock to check if all updates received
         if self._round_updates_lock.locked():
             self._round_updates_lock.release_async()
+        
+        self._notification = False
 
     async def _check_updates_already_received(self):
         for se in self._sources_expected:
@@ -92,24 +89,7 @@ class DFLUpdateHandler(UpdateHandler):
                     logging.info(f"Update already received from source: {se} | ({len(self._sources_received)}/{len(self._sources_expected)}) Updates received")
                     self._sources_received.add(se)
         
-    async def storage_update(self, model, weight, source, round):
-        """
-            Stores an update in the update queue if it has not been previously received for the same source and round.
-
-            This method ensures that only one update per source and round is stored, avoiding duplicates.
-            If all expected updates for the current round have been received, it triggers the `all_updates_received()` method.
-
-            Args:
-                model: The model associated with the update.
-                weight: The weight assigned to the update.
-                source (str): The source identifier of the update.
-                round (int): The training round in which the update was received.
-
-            Logs:
-                - Stores and logs the update if it's new for the round.
-                - Logs a duplicate update if an identical one already exists.
-                - Logs missing sources if not all expected updates have been received.
-        """
+    async def storage_update(self, model, weight, source, round, local=False):
         time_received = time.time()
         if source in self._sources_expected:
             updt = Update(model, weight, source, round, time_received)
@@ -126,30 +106,19 @@ class DFLUpdateHandler(UpdateHandler):
                 updates_left = self._sources_expected.difference(self._sources_received)
                 logging.info(f"Updates received ({len(self._sources_received)}/{len(self._sources_expected)}) | Missing nodes: {updates_left}")
                 if self._round_updates_lock.locked() and not updates_left:
-                    await self._all_updates_received()
+                    all_rec = await self._all_updates_received()
+                    if all_rec:
+                        await self._notify()
             await self._updates_storage_lock.release_async()
         else:
-            logging.info(f"Discard update | source: {source} not in expected updates for this Round")
+            if not source in self._sources_received:
+                logging.info(f"Discard update | source: {source} not in expected updates for this Round")
             
     async def get_round_updates(self):
-        """
-            Retrieves the latest updates received in the current round.
-
-            This method collects updates from all received sources, prioritizing the most recent update
-            stored in the queue. If an expected update is missing, it attempts to use the last received 
-            update from that source instead.
-
-            Returns:
-                dict: A dictionary mapping each source to a tuple `(model, weight)`, containing 
-                    the most recent update for that source.
-
-            Logs:
-                - Logs missing sources if expected updates have not been received.
-                - Logs when a missing update is replaced by the last received update.
-        """
         await self._updates_storage_lock.acquire_async()
         updates_missing = self._sources_expected.difference(self._sources_received)
         if updates_missing:
+            self._missing_ones = updates_missing
             logging.info(f"Missing updates from sources: {updates_missing}")
         updates = {}
         for sr in self._sources_received:
@@ -157,7 +126,7 @@ class DFLUpdateHandler(UpdateHandler):
             last_updt_received = self.us[sr][0]
             updt: Update = None
             updt = source_historic[-1] # Get last update received
-            if last_updt_received == updt:
+            if last_updt_received and last_updt_received == updt:
                 logging.info(f"Missing update source: {sr}, using last update received..")
             else:
                 last_updt_received = updt
@@ -190,24 +159,45 @@ class DFLUpdateHandler(UpdateHandler):
         await self._updates_storage_lock.release_async()
                   
     async def get_round_missing_nodes(self):
-        await self._updates_storage_lock.acquire_async()
-        updates_left = self._sources_expected.difference(self._sources_received)
-        await self._updates_storage_lock.release_async()
-        return updates_left
+        # await self._updates_storage_lock.acquire_async()
+        # updates_left = self._sources_expected.difference(self._sources_received)
+        # await self._updates_storage_lock.release_async()
+        return self._missing_ones
     
     async def notify_if_all_updates_received(self):
         logging.info("Set notification when all expected updates received")
         await self._round_updates_lock.acquire_async()
-
+        await self._updates_storage_lock.acquire_async()
+        all_received = await self._all_updates_received()
+        await self._updates_storage_lock.release_async()
+        if all_received:
+            await self._notify()
+            
+            
     async def stop_notifying_updates(self):
-        logging.info("Stop notifications updates")
         if self._round_updates_lock.locked():
+            logging.info("Stop notification updates")
             await self._round_updates_lock.release_async()
+            
+    async def _notify(self):
+        await self._notification_sent_lock.acquire_async()
+        if self._notification:
+            await self._notification_sent_lock.release_async()
+            return
+        self._notification = True
+        await self.stop_notifying_updates()
+        await self._notification_sent_lock.release_async()
+        logging.info("🔄 Notifying aggregator to release aggregation")
+        await self.agg.notify_all_updates_received()
+        
     
     async def _all_updates_received(self):
         updates_left = self._sources_expected.difference(self._sources_received)
+        all_received = False
         if len(updates_left) == 0:
-            logging.info("All updates have been received this round | releasing aggregation lock")
+            logging.info("All updates have been received this round")
             await self._round_updates_lock.release_async()
-            #await self.agg.notify_all_updates_received()
+            all_received = True          
+        return all_received
+
