+++ nebula/core/topologymanagement/candidateselection/hetcandidateselector.py
@@ -1,4 +1,4 @@
-from nebula.core.neighbormanagement.candidateselection.candidateselector import CandidateSelector
+from nebula.core.topologymanagement.candidateselection.candidateselector import CandidateSelector
 from nebula.core.utils.locker import Locker
 
 class HETCandidateSelector(CandidateSelector):
