+++ nebula/core/topologymanagement/neighborpolicies/ringneighborpolicy.py
@@ -1,23 +1,20 @@
-from nebula.core.neighbormanagement.neighborpolicies.neighborpolicy import NeighborPolicy
+from nebula.core.topologymanagement.neighborpolicies.neighborpolicy import NeighborPolicy
 from nebula.core.utils.locker import Locker
+import random
 
-class FCNeighborPolicy(NeighborPolicy):
-    
+class RINGNeighborPolicy(NeighborPolicy):
+        
     def __init__(self):
-        self.max_neighbors = None
+        self.max_neighbors = 2
         self.nodes_known = set()
         self.neighbors = set()
-        self.addr = None
         self.neighbors_lock = Locker(name="neighbors_lock")
         self.nodes_known_lock = Locker(name="nodes_known_lock")
+        self.addr = ""
         
     def need_more_neighbors(self):
-        """
-            Fully connected network requires to be connected to all devices, therefore,
-            if there are more nodes known that self.neighbors, more neighbors are required
-        """
         self.neighbors_lock.acquire()
-        need_more = (len(self.neighbors) < len(self.nodes_known))
+        need_more = len(self.neighbors) < self.max_neighbors
         self.neighbors_lock.release()
         return need_more
     
@@ -26,50 +23,33 @@ class FCNeighborPolicy(NeighborPolicy):
         Args:
             config[0] -> list of self neighbors
             config[1] -> list of nodes known on federation
-            config[2] -> self addr
-            config[3] -> NodeManager reference
+            config[2] -> self.addr
         """
         self.neighbors_lock.acquire()
-        self.neighbors = config[0] 
+        self.neighbors = config[0]
         self.neighbors_lock.release()
         for addr in config[1]:
                 self.nodes_known.add(addr)
-        self.addr
+        self.addr = config[2]
             
     def accept_connection(self, source, joining=False):
         """
             return true if connection is accepted
         """
+        ac = False
         self.neighbors_lock.acquire()
-        ac = (not source in self.neighbors)
+        if not joining:    
+            ac = not source in self.neighbors
+        else:
+            ac = not len(self.neighbors) == self.max_neighbors
         self.neighbors_lock.release()
         return ac
     
     def meet_node(self, node):
-        """
-            Update the list of nodes known on federation
-        """
         self.nodes_known_lock.acquire()
-        if node != self.addr:
-            self.nodes_known.add(node)
+        self.nodes_known.add(node)
         self.nodes_known_lock.release()
         
-    def get_nodes_known(self, neighbors_too=False, neighbors_only=False):     
-        if neighbors_only:
-            self.neighbors_lock.acquire()
-            no = self.neighbors.copy()
-            self.neighbors_lock.release()
-            return no
-        
-        self.nodes_known_lock.acquire()
-        nk = self.nodes_known.copy()
-        if not neighbors_too:
-            self.neighbors_lock.acquire()
-            nk = self.nodes_known - self.neighbors
-            self.neighbors_lock.release()
-        self.nodes_known_lock.release()
-        return nk     
-    
     def forget_nodes(self, node, forget_all=False):
         self.nodes_known_lock.acquire()
         if forget_all:
@@ -78,34 +58,38 @@ class FCNeighborPolicy(NeighborPolicy):
             self.nodes_known.discard(node)
         self.nodes_known_lock.release()
         
+    def get_nodes_known(self, neighbors_too=False, neighbors_only=False):
+        self.nodes_known_lock.acquire()
+        nk = self.nodes_known.copy()
+        if not neighbors_too:
+            self.neighbors_lock.acquire()
+            nk = self.nodes_known - self.neighbors 
+            self.neighbors_lock.release()
+        self.nodes_known_lock.release()
+        return nk 
+        
     def get_actions(self): 
         """
             return list of actions to do in response to connection
                 - First list represents addrs argument to LinkMessage to connect to
                 - Second one represents the same but for disconnect from LinkMessage
         """ 
-        return [self._connect_to(), self._disconnect_from()]
-          
-    
-    def _disconnect_from(self):
-        return ""
-    
-    def _connect_to(self):
-        ct = ""
         self.neighbors_lock.acquire()
-        ct = " ".join(self.neighbors)
+        ct_actions = []
+        df_actions = []
+        if len(self.neighbors) < self.max_neighbors:
+            list_neighbors = list(self.neighbors)
+            index = random.randint(0, len(list_neighbors)-1)
+            node = list_neighbors[index]
+            ct_actions.append(node)                            # connect to
+            df_actions.append(self.addr)                       # disconnect from
         self.neighbors_lock.release()
-        return ct
+        return [ct_actions, df_actions]
     
     def update_neighbors(self, node, remove=False):
-        if node == self.addr:
-            return
         self.neighbors_lock.acquire()
         if remove:
-            try:
-                self.neighbors.remove(node)
-            except KeyError:
-                pass    
+            self.neighbors.remove(node)
         else:
             self.neighbors.add(node)
         self.neighbors_lock.release()
